+++ src/persist/entities/EntityUpload.ts
@@ -1,19 +1,7 @@
-import { UploadState } from './../../sync/task/UploaderTask';
-import { S3StreamSessionDetails } from 'upaki-cli';
 import * as crypto from 'crypto';
 import { Database } from '../Database';
-
-export interface EntityUploadData {
-    key: string;
-    lastModifies: string;
-    file_id: string;
-    folder_id: string;
-    sessionData?: S3StreamSessionDetails;
-    state: UploadState;
-    Etag: string;
-    path: string;
-}
-
+import { EntityUploadData } from '../../api/entity';
+import { Environment } from '../../config/env';
 
 export class EntityUpload {
     private static _instance: EntityUpload;
@@ -33,16 +21,16 @@ export class EntityUpload {
     save(upload: EntityUploadData, callback: (err: Error, data: any) => void) {
         let key = this.MD5SRC(upload.path);
         let data = JSON.stringify(upload);
-        Database.Instance.All(`SELECT data FROM ${this.table} WHERE key=?`, [key], (err, row) => {
+        Database.Instance.All(`SELECT data FROM ${this.table} WHERE key=? and user_id=?`, [key, Environment.config.credentials.userId], (err, row) => {
             if (err) {
                 callback(err, undefined);
             } else {
                 if (row[0]) {
-                    Database.Instance.Run(`UPDATE ${this.table} SET data=? WHERE key = ?`, [data, key], (errInsert) => {
+                    Database.Instance.Run(`UPDATE ${this.table} SET data=? WHERE key = ? and user_id=?`, [data, key, Environment.config.credentials.userId], (errInsert) => {
                         callback(errInsert, undefined);
                     });
                 } else {
-                    Database.Instance.Run(`INSERT INTO ${this.table}(key, data) VALUES(?,?)`, [key, data], (errInsert) => {
+                    Database.Instance.Run(`INSERT INTO ${this.table}(key, data, user_id) VALUES(?,?,?)`, [key, data, Environment.config.credentials.userId], (errInsert) => {
                         callback(errInsert, undefined);
                     });
                 }
@@ -53,7 +41,7 @@ export class EntityUpload {
     getFile(path: string): Promise<EntityUploadData> {
         return new Promise<EntityUploadData>((resolve, reject) => {
             let key = this.MD5SRC(path);
-            Database.Instance.Get(`SELECT data FROM ${this.table} WHERE key=?`, [key], (err, row) => {
+            Database.Instance.Get(`SELECT data FROM ${this.table} WHERE key=? and user_id=?`, [key, Environment.config.credentials.userId], (err, row) => {
                 if (err) {
                     reject(err);
                 } else {
@@ -73,7 +61,7 @@ export class EntityUpload {
                 return this.MD5SRC(el);
             });
             // console.log(keys.join(','));
-            Database.Instance.All(`SELECT key, data FROM ${this.table} WHERE key in(${keys.map(k => { return `'${k}'`; }).join(`,`)})`, [], (err, row) => {
+            Database.Instance.All(`SELECT key, data FROM ${this.table} WHERE key in(${keys.map(k => { return `'${k}'`; }).join(`,`)}) and user_id=?`, [Environment.config.credentials.userId], (err, row) => {
                 if (err) {
                     reject(err);
                 } else {
@@ -96,7 +84,7 @@ export class EntityUpload {
         return new Promise<number>((resolve, reject) => {
             let key = this.MD5SRC(oldPath);
             let newKey = this.MD5SRC(newPath);
-            Database.Instance.Get(`SELECT data FROM ${this.table} WHERE key=?`, [key], (err, row) => {
+            Database.Instance.Get(`SELECT data FROM ${this.table} WHERE key=? and user_id=?`, [key, Environment.config.credentials.userId], (err, row) => {
                 if (err) {
                     reject(err);
                     return;
@@ -105,7 +93,7 @@ export class EntityUpload {
                     let data = JSON.parse(row.data) as EntityUploadData;
                     data.path = newPath;
 
-                    Database.Instance.Run(`UPDATE ${this.table} SET key=?, data=? WHERE key = ?`, [newKey, JSON.stringify(data), key], (errInsert) => {
+                    Database.Instance.Run(`UPDATE ${this.table} SET key=?, data=? WHERE key = ? and user_id=?`, [newKey, JSON.stringify(data), key, Environment.config.credentials.userId], (errInsert) => {
                         if (errInsert) {
                             reject(errInsert);
                             return;
@@ -122,7 +110,7 @@ export class EntityUpload {
     delete(path: string): Promise<number> {
         return new Promise<number>((resolve, reject) => {
             let key = this.MD5SRC(path);
-            Database.Instance.Run(`DELETE FROM ${this.table} WHERE key = ?`, [key], (err) => {
+            Database.Instance.Run(`DELETE FROM ${this.table} WHERE key = ? and user_id=?`, [key, Environment.config.credentials.userId], (err) => {
                 if (err) {
                     reject(err);
                 } else {
